[KSF-19] Add Product Category Config on admin dashboard and change route name
