+++ src/app/admin/dashboard/[store_id]/page.tsx
