+++ src/app/admin/dashboard/page.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 
-import AdminDashboard from "@/components/admin/dashboard/aDashboard";
+import AdminDashboard from "@/components/admin/dashboard/adminDashboard";
 
 type Props = {};
